sumo_liveUpdate_ui
This compressed package provides a UI that can run SUMO in real-time and record the number of runs and objective function values. It requires you to provide SUMO's road network file, traffic flow file, and SUMO configuration file, and click the optimization button to run it.
You can pause the optimization at any time and display the final result in the ListBox in the bottom right corner. Then, an iteration message will pop up, and the system preset maximum number of iterations is 30.
This UI is running on a Sqlite3 database, and Python version 3.6 and above has been validated by the author as valid.